+++ miracle/__init__.py
@@ -1,1 +1,1 @@
-VERSION = '1.1.5.dev0'
+VERSION = '1.1.5'